svelte(phase 3): absorb the four chat selectors into ChatComposer
Move model-selector / thinking-selector / slash-command / mention-autocomplete
(~808 LOC) into ChatComposer.svelte's <script module> as exported functions.
runChatComposer's selector params now default to the absorbed setups, so
production uses them while the runner's unit tests still inject mocks. The pure
chat-selectors.js helpers stay a module.

The four modules are deleted; their test files move to components/session/ with
the import path repointed to ./ChatComposer.svelte (all 72 cases green).

Fix a recursion bug the e2e caught: the runner's internal thinking wrapper was
named setupThinkingLevelSelector (identical to the absorbed export), so the
param default { setupThinkingLevelSelector } bound to the wrapper and recursed.
Renamed it to loadThinkingSelector, matching the other load* wrappers.
