+++ docs/dev/svelte-migration-plan.md
@@ -609,6 +609,7 @@ per increment.
 
 | Status | Item |
 |---|---|
+| ✅ DONE + verified | **Four chat selectors → `<ChatComposer>` `<script module>`.** Absorbed `model-selector.js`, `thinking-selector.js`, `slash-command.js`, `mention-autocomplete.js` (≈808 LOC) into ChatComposer's module script as exported functions. `runChatComposer`'s selector params now **default** to the absorbed setups (`modelSelector = { setupModelSelector }`, etc.) so production uses them while the runner's tests still inject mocks. The pure `chat-selectors.js` helpers stay a module (imported by the absorbed code). **4 modules DELETED**; their 4 test files moved to `components/session/` with the import path repointed to `./ChatComposer.svelte` — all 54 selector cases + 18 runner cases green. **Bug caught by e2e:** the runner's internal thinking wrapper was named `setupThinkingLevelSelector` — identical to the absorbed export — so the param default `{ setupThinkingLevelSelector }` bound to the wrapper → infinite recursion (slash/mention setup never ran). Fixed by renaming the wrapper to `loadThinkingSelector` (matching `loadModelSelector`/`loadSlashSelector`/`loadMentionSelector`). Verified: web 527 + knip clean + build (no a11y) + `go test ./internal/ui` + go vet + Desktop-Chrome e2e (slash/mention/composer-keyboard/chat/context-usage green; full **57/2**). |
 | ✅ DONE + verified | **`chat-composer-runner.js` (1121 LOC) → `<ChatComposer>`.** Moved `runChatComposer` verbatim into `ChatComposer.svelte`'s `<script module>` (exported, keeping its full DI signature); the instance `<script>` `onMount` calls it exactly as before. The model/thinking/slash/mention selectors + `chat-api` stay as injected helper modules (passed in). Module-scope `icon`/`t`/`X`/`TextQuote` feed the runner; the markup reuses them. **`chat-composer-runner.js` DELETED**; its 586-LOC / 18-case behavioural suite moved to `components/session/ChatComposer.test.js` with only the import path repointed to `./ChatComposer.svelte` (the `<script module>` export) — **zero test rewrites**, all 18 still green. Verified: web 527 + knip clean + build (no a11y) + autofixer clean + `go test ./internal/ui` + go vet + Desktop-Chrome e2e (chat/slash/mention/context-usage/composer-keyboard/btw green; full **57/2**). |
 | ✅ DONE + verified | **`live-reload-runner.js` → `<LiveReload>` (reactive-only).** Absorbed the 371-LOC runner into `LiveReload.svelte`'s `onMount`: SSE connect/reconnect/backoff, the follow/scroll state machine, the streaming chat-preview wiring, stats/title updates, and the new-entry highlight (`highlightNewEntry` inlined). Hardcoded reactive mode — the dead non-reactive DOM-patching branch (`appendEntry`/`upsertEntry`/`refreshEntriesAffectedByToolResult`/`createSeenEntrySet`) is gone; reload reconciles purely via `window.__piReconcileEntries` and tracks new ids for follow/scroll. The streaming preview's `renderMarkdown` now uses the shared `safeMarkedParse` (globally-configured/sanitized marked) with an `escapeHtml` fallback, so the parallel renderer dies. Added an unmount disposer (closes the EventSource, clears the timer, removes listeners). **DELETED `live-reload-runner.js`, `live-renderer.js`, `live-entries.js` (+ their tests)** — the second imperative entry renderer + the imperative `#messages` patchers. **Kept** as live helpers: `live-events.js` (SSE/reload primitives), `live-scroll.js`, `live-stats.js`, `chat-preview.js`. Go guards updated: the two `export_html` source guards repoint runner→`LiveReload.svelte`; `TestLiveReloadUpdatesExistingAssistantWhenToolResultsArrive` removed (the reactive model + `<SessionContent>` now provide that guarantee); `TestLiveReloadEntriesInheritCurrentToggleState` trimmed to the shared hook; the renderer-markup guard repoints to `session-entry-renderer.js`. Verified: web 527 + knip clean + build (no a11y) + autofixer clean + `go test ./internal/ui` + go vet + Desktop-Chrome e2e (live-reload/chat/btw/context-usage green; full **57/2**). |
 | ✅ DONE + verified | **`GitFooter.svelte`** (chat tranche, first cut). Extracted the git bar (branch indicator + smart PR split button + branch rename) from `git-footer.js` (264 LOC, called directly in `ChatComposer.onMount`) into a child `<GitFooter {sessionId} />` that owns the `.pi-git-bar` markup — including the always-available `#pi-btw-button` (still wired by `<BtwPopup>`). `git-api.js` stays as a **pure** fetch helper (injectable `gitApi` prop, default = the real module). Dropped the dead **dataset-sync first-paint path** (the SPA never stamps `data-git-*`, so it was inert — the bar is driven by the async `getGitInfo`). **`chat/git-footer.js` + `git-footer.test.js` DELETED** (ported to `GitFooter.test.js`, 8 cases). Verified: web 534 + knip clean + no a11y (only the standard `{@html icon}` notes) + `go test ./internal/ui` + go vet + Desktop-Chrome e2e (chat/btw/composer/slash/mention green; full 57/2). |
